Require Fog to be enabled before using effect
In Fog is not enabled, there are no valid shader variants that can be
selected from.  In the Editor, Unity will silently choose another
variant, and not mention that no valid variants are usable by the
effect.  However, when you generate a stand-alone build, visual
corruption will happen due to the FadeToSkybox.shader not being used.

We can check for Fog being enabled, and also print out an error in the
Component dialog box.  Additionally, if Fog is not enabled, the shader